pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anh Phan Le Nhat <phanlerda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.13"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
python-dotenv = "^1.1.1"
pydantic-settings = "^2.10.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.41"}
psycopg2-binary = "^2.9.10"
alembic = "^1.16.2"
python-multipart = "^0.0.20"
sentence-transformers = "^4.1.0"
pymupdf = "^1.26.1"
unstructured = {extras = ["pdf"], version = "^0.17.2"}
langchain = "^0.3.26"
tiktoken = "^0.9.0"
qdrant-client = ">=1.9.0,<1.10.0"
tavily-python = "^0.7.8"
ragas = "^0.2.15"
pandas = ">=2.0.0"
langchain-google-genai = "^2.1.5"
langchain-core = "^0.3.66"
langchain-community = "^0.3.26"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.5.0"
pydantic = {extras = ["email"], version = "^2.11.7"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
